Add NLD_NUMERIC_CHARACTER to TO_CHAR
As reported in DZ_JSON, systesm with NLS_NUMERIC__CHARACTER set to use
the comma as the decimal separater will generate problematic numeric
values with my libraries.  For JSON the answer is easy as JSON does not
allow localization.  For WKT its trickier as the specification does
indeed seem to allow localization.  However, it appears in general that
support for such localized WKT is tepid as best.  I am just following
the crowd by not supporting localized WKT.